+++ backend/pkg/kafka/service.go
@@ -117,6 +117,14 @@ func (s *Service) keepAlive() {
 				log.Warn("heartbeat: lost connection to broker", zap.Error(err), zap.String("broker", broker.Addr()), zap.Int32("id", broker.ID()))
 				_ = broker.Close()
 				_ = broker.Open(s.Client.Config())
+
+				log.Info("trying to refresh cluster metadata after we've lost connection to at least one broker")
+				metadataErr := s.Client.RefreshMetadata()
+				if metadataErr == nil {
+					log.Info("refreshed cluster metadata successfully")
+				} else {
+					log.Warn("failed to refresh cluster metadata after we've lost connection to at least one broker")
+				}
 				continue
 			}