+++ frontend/src/state/backendApi.ts
@@ -13,7 +13,7 @@ import { appGlobal } from "./appGlobal";
 import { uiState } from "./uiState";
 import { notification } from "antd";
 
-const REST_TIMEOUT_SEC = IsDev ? 5 : 25;
+const REST_TIMEOUT_SEC = 25;
 export const REST_CACHE_DURATION_SEC = 20;
 const REST_DEBUG_BASE_URL = null// || "http://localhost:9090"; // only uncommented using "npm run build && serve -s build"